--- app/demo/page.tsx
@@ -1,160 +1,0 @@
-'use client'
-
-import { zodResolver } from '@hookform/resolvers/zod'
-import { KeyRound, Mail } from 'lucide-react'
-import { useRouter } from 'next/navigation'
-import { useState } from 'react'
-import type { SubmitHandler } from 'react-hook-form'
-import { useForm } from 'react-hook-form'
-import { z } from 'zod'
-import { Alert, AlertDescription } from '@/components/ui/alert'
-import { Button } from '@/components/ui/button'
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card'
-import { Input } from '@/components/ui/input'
-import { Label } from '@/components/ui/label'
-import { createClient } from '@/src/client/supabase'
-
-// バリデーションスキーマ
-const loginSchema = z.object({
-  email: z
-    .string()
-    .min(1, 'メールアドレスは必須です')
-    .email('有効なメールアドレスを入力してください'),
-  password: z.string().min(6, 'パスワードは6文字以上で入力してください'),
-})
-
-type LoginFormValues = z.infer<typeof loginSchema>
-
-export default function LoginDemo() {
-  const [error, setError] = useState<string | null>(null)
-  const [loading, setLoading] = useState(false)
-  const router = useRouter()
-  const [supabase] = useState(() => createClient())
-
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-  } = useForm<LoginFormValues>({
-    resolver: zodResolver(loginSchema),
-  })
-
-  const onSubmit: SubmitHandler<LoginFormValues> = async (data) => {
-    setError(null)
-    setLoading(true)
-
-    try {
-      const { error } = await supabase.auth.signInWithPassword({
-        email: data.email,
-        password: data.password,
-      })
-
-      if (error) {
-        setError(error.message)
-        return
-      }
-
-      // ログイン成功
-      router.push('/')
-      router.refresh()
-    } catch {
-      setError('ログイン中にエラーが発生しました')
-    } finally {
-      setLoading(false)
-    }
-  }
-
-  return (
-    <div className="flex min-h-screen items-center justify-center bg-background p-4">
-      <div className="w-full max-w-md space-y-6">
-        {/* ヘッダー */}
-        <div className="text-center space-y-2">
-          <div className="inline-flex items-center justify-center w-16 h-16 rounded-full bg-primary mb-3">
-            <KeyRound className="w-8 h-8 text-primary-foreground" />
-          </div>
-          <h1 className="text-3xl font-bold text-foreground">Kojiro Mahjong App</h1>
-          <p className="text-muted-foreground">麻雀リーグ管理システム</p>
-        </div>
-
-        {/* ログインカード */}
-        <Card className="border-border shadow-lg">
-          <CardHeader className="space-y-2">
-            <CardTitle className="text-2xl text-foreground">ログイン</CardTitle>
-            <CardDescription className="text-muted-foreground">
-              メールアドレスとパスワードを入力してください
-            </CardDescription>
-          </CardHeader>
-          <CardContent>
-            <form onSubmit={handleSubmit(onSubmit)} className="space-y-4">
-              {/* メールアドレス */}
-              <div className="space-y-2">
-                <Label htmlFor="email" className="text-sm font-medium text-foreground">
-                  メールアドレス
-                </Label>
-                <div className="relative">
-                  <Mail className="absolute left-3 top-1/2 -translate-y-1/2 h-4 w-4 text-muted-foreground" />
-                  <Input
-                    id="email"
-                    type="email"
-                    placeholder="example@example.com"
-                    className="pl-10 h-11"
-                    {...register('email')}
-                    disabled={loading}
-                  />
-                </div>
-                {errors.email && (
-                  <p className="text-sm text-destructive mt-1">{errors.email.message}</p>
-                )}
-              </div>
-
-              {/* パスワード */}
-              <div className="space-y-2">
-                <Label htmlFor="password" className="text-sm font-medium text-foreground">
-                  パスワード
-                </Label>
-                <div className="relative">
-                  <KeyRound className="absolute left-3 top-1/2 -translate-y-1/2 h-4 w-4 text-muted-foreground" />
-                  <Input
-                    id="password"
-                    type="password"
-                    placeholder="••••••••"
-                    className="pl-10 h-11"
-                    {...register('password')}
-                    disabled={loading}
-                  />
-                </div>
-                {errors.password && (
-                  <p className="text-sm text-destructive mt-1">{errors.password.message}</p>
-                )}
-              </div>
-
-              {/* エラーメッセージ */}
-              {error && (
-                <Alert variant="destructive">
-                  <AlertDescription>{error}</AlertDescription>
-                </Alert>
-              )}
-
-              {/* 送信ボタン */}
-              <Button type="submit" className="w-full h-11 font-medium" disabled={loading}>
-                {loading ? 'ログイン中...' : 'ログイン'}
-              </Button>
-            </form>
-
-            {/* フッター */}
-            <div className="mt-6 pt-4 border-t border-border">
-              <p className="text-sm text-center text-muted-foreground">
-                アカウントをお持ちでない方は管理者にお問い合わせください
-              </p>
-            </div>
-          </CardContent>
-        </Card>
-
-        {/* バージョン情報 */}
-        <p className="text-xs text-center text-muted-foreground">
-          Version 1.0.0 | Powered by Next.js 15 & Supabase
-        </p>
-      </div>
-    </div>
-  )
-}
