+++ app/page.tsx
@@ -1,9 +1,11 @@
 'use client'
 
 import type { User } from '@supabase/supabase-js'
+import { BarChart3, LogOut, Users } from 'lucide-react'
 import { useRouter } from 'next/navigation'
 import { useEffect, useState } from 'react'
 import { Button } from '@/components/ui/button'
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card'
 import { createClient } from '@/src/client/supabase'
 
 export default function Home() {
@@ -39,27 +41,64 @@ export default function Home() {
     router.refresh()
   }
 
+  // 未ログイン状態
   if (!user) {
     return (
-      <div style={{ padding: '20px' }}>
-        <p>ログインしていません</p>
-        <button type="button" onClick={() => router.push('/login')}>
-          ログイン
-        </button>
-        <Button variant="outline" size="icon" aria-label="Submit">
-          てすとです
-        </Button>
+      <div className="flex min-h-screen items-center justify-center bg-background p-4">
+        <Card className="w-full max-w-md border-border shadow-lg">
+          <CardHeader className="text-center space-y-2">
+            <div className="inline-flex items-center justify-center w-16 h-16 rounded-full bg-muted mx-auto mb-3">
+              <Users className="w-8 h-8 text-muted-foreground" />
+            </div>
+            <CardTitle className="text-2xl">ログインが必要です</CardTitle>
+            <CardDescription>このページを表示するにはログインしてください</CardDescription>
+          </CardHeader>
+          <CardContent className="space-y-3">
+            <Button onClick={() => router.push('/login')} className="w-full h-11 font-medium">
+              ログインページへ
+            </Button>
+            <p className="text-xs text-center text-muted-foreground">
+              アカウントをお持ちでない方は管理者にお問い合わせください
+            </p>
+          </CardContent>
+        </Card>
       </div>
     )
   }
 
+  // ログイン済み状態
   return (
-    <div style={{ padding: '20px' }}>
-      <h1>Kojiro Mahjong App</h1>
-      <p>ようこそ、{user.email} さん</p>
-      <button type="button" onClick={handleLogout}>
-        ログアウト
-      </button>
+    <div className="min-h-screen bg-background">
+      {/* ヘッダー */}
+      <header className="sticky top-0 z-50 w-full border-b border-border bg-background/95 backdrop-blur supports-[backdrop-filter]:bg-background/60">
+        <div className="container flex h-16 items-center justify-between px-4">
+          <div className="flex items-center gap-3">
+            <div className="flex items-center justify-center w-10 h-10 rounded-lg bg-primary">
+              <BarChart3 className="w-6 h-6 text-primary-foreground" />
+            </div>
+            <div>
+              <h1 className="text-xl font-bold text-foreground">Kojiro Mahjong App</h1>
+              <p className="text-xs text-muted-foreground">麻雀リーグ管理システム</p>
+            </div>
+          </div>
+          <Button variant="outline" onClick={handleLogout} className="flex items-center gap-2">
+            <LogOut className="w-4 h-4" />
+            ログアウト
+          </Button>
+        </div>
+      </header>
+
+      {/* メインコンテンツ */}
+      <main className="container px-4 py-8">
+        <div className="flex items-center justify-center min-h-[60vh]">
+          <Card className="w-full max-w-md border-border shadow-lg">
+            <CardHeader className="text-center space-y-2">
+              <CardTitle className="text-2xl">ようこそ</CardTitle>
+              <CardDescription>{user.email}</CardDescription>
+            </CardHeader>
+          </Card>
+        </div>
+      </main>
     </div>
   )
 }
